convert grid and block idx to 32 bits

+++ src/muon/core.rs
@@ -52,7 +52,7 @@ impl MuonCore {
         self.scheduler.spawn_single_warp()
     }
 
-    pub fn spawn_n_warps(&mut self, pc: u32, block_idx: (u16, u16, u16), thread_idxs: Vec<Vec<(u16, u16, u16)>>) {
+    pub fn spawn_n_warps(&mut self, pc: u32, block_idx: (u32, u32, u32), thread_idxs: Vec<Vec<(u32, u32, u32)>>) {
         self.scheduler.spawn_n_warps(pc, thread_idxs.len());
         self.warps.iter_mut().zip(thread_idxs.iter()).for_each(|(warp, warp_thread_idxs)| {
             warp.set_block_threads(block_idx, warp_thread_idxs);
